+++ effekt/shared/src/main/scala/effekt/generator/llvm/Tree.scala
@@ -1,4 +1,5 @@
 package effekt
+package generator
 package llvm
